newdevs/requirements.txt
asgiref==3.2.3
Django==2.2
django-extensions==2.2.5
django-generate-secret-key==1.0.2
django-material==1.6.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
Pillow==6.2.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
